Split SocketManager.
