+++ src/network/SocketManagerTimeouts.cpp
@@ -1,0 +1,105 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   SocketManagerTimeouts.cpp                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/08 15:03:19 by irychkov          #+#    #+#             */
+/*   Updated: 2025/06/08 18:58:11 by irychkov         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "http/HttpResponse.hpp"     // for HttpResponse
+#include "http/handleCgi.hpp"        // for CgiProcess
+#include "network/SocketManager.hpp" // for ClientInfo, SocketManager, TIMEOUT
+#include "utils/Logger.hpp"          // for LogLevel, Logger
+#include "utils/filesystemUtils.hpp" // for getCurrentTime
+#include <map>                       // for map
+#include <optional>                  // for optional
+#include <poll.h>                    // for pollfd, POLLIN
+#include <queue>                     // for queue
+#include <string>                    // for operator+, allocator, to_string
+#include <time.h>                    // for time_t, size_t
+#include <vector>                    // for vector
+
+bool SocketManager::isHeaderTimeout(int fd, time_t now) {
+    ClientInfo& client = _client_info[fd];
+
+    if (client.responses.empty() && client.current_raw_response.empty() && !client.headerComplete &&
+        client.headerBytesReceived > 0 &&
+        now - client.connectionStartTime > HEADER_TIMEOUT_SECONDS) {
+        Logger::logFrom(LogLevel::WARN, "SocketManager",
+                        "Header timeout on fd: " + std::to_string(fd));
+        respondError(fd, 408);
+        return true;
+    }
+
+    return false;
+}
+
+bool SocketManager::isBodyTimeout(int fd, time_t now) {
+    ClientInfo& client = _client_info[fd];
+    if (client.headerComplete && now - client.connectionStartTime > TIMEOUT) {
+        Logger::logFrom(LogLevel::WARN, "SocketManager",
+                        "Body timeout on fd: " + std::to_string(fd));
+        respondError(fd, 408);
+        return true;
+    }
+    return false;
+}
+
+bool SocketManager::isSendTimeout(int fd, time_t now) {
+    ClientInfo& client = _client_info[fd];
+    if (!client.responses.empty() && !client.current_raw_response.empty() &&
+        now - client.lastSendAttemptTime > TIMEOUT) {
+        Logger::logFrom(LogLevel::WARN, "SocketManager",
+                        "Send timeout on fd: " + std::to_string(fd));
+        return true;
+    }
+    return false;
+}
+
+bool SocketManager::isIdleTimeout(int fd, time_t now) {
+    ClientInfo& client = _client_info[fd];
+    if (client.cgiProcess.has_value()) {
+        return false;
+    }
+    if (client.responses.empty() && client.current_raw_response.empty() && !client.headerComplete &&
+        client.headerBytesReceived == 0 && now - client.lastRequestTime > TIMEOUT) {
+        Logger::logFrom(LogLevel::WARN, "SocketManager",
+                        "Idle timeout on fd: " + std::to_string(fd));
+        return true;
+    }
+    return false;
+}
+
+bool SocketManager::checkClientTimeouts(int client_fd, size_t index) {
+    if (!_client_info.count(client_fd))
+        return false;
+
+    time_t now = getCurrentTime();
+    if (isIdleTimeout(client_fd, now) || isSendTimeout(client_fd, now)) {
+        cleanupClientConnectionClose(client_fd, index);
+        return false;
+    }
+    if (isHeaderTimeout(client_fd, now) || isBodyTimeout(client_fd, now)) {
+        _poll_fds[index].events &= ~POLLIN;
+        return true;
+    }
+    return false;
+}
+
+bool SocketManager::checkRequestLimits(int fd) {
+    ClientInfo& client = _client_info[fd];
+
+    // Only enforce header-length limit while headers are still incomplete
+    if (client.headerBytesReceived > HEADER_MAX_LENGTH) {
+        Logger::logFrom(LogLevel::WARN, "SocketManager",
+                        "Request header too large from fd: " + std::to_string(fd));
+        respondError(fd, 431); // Request Header Fields Too Large
+        return true;
+    }
+
+    return false;
+}
